# perf(lib): faster ray-triangle intersection tests