--- a/docs/source/conf.py
+++ b/docs/source/conf.py
@@ -147,6 +147,7 @@ myst_heading_anchors = 3
 
 myst_enable_extensions = [
     "amsmath",
+    "attrs_inline",
     "colon_fence",
     "dollarmath",
     "html_admonition",