Triangles / rays are grouped by `batch_size` to (possibly) accelerate the scan procedure. Moreover, `jax.lax.dynamic_slice_in_dim` is used to avoid having to reshape the arrays the have the `batched` dimension as the first dimension (required by `jax.lax.scan`).

See https://github.com/jax-ml/jax/discussions/30470

---

*Comments*

For an unknown reason (maybe the machine on which CI runs), the CodSpeed benchmarks report a regression for some tests, while I cannot reproduce this locally. All my tests show a significant improvement for related benchmarks. See timings on my machine:

**Before**

```
                                   Benchmark Results                                   
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃                        Benchmark ┃     Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│                test_image_method │        87,736ns │       16.8% │    3.00s │ 8,850 │
│                      test_fermat │    86,295,515ns │        2.9% │    2.90s │    32 │
│ test_rays_intersect_any_triangle │   168,445,920ns │        7.6% │    2.62s │    14 │
│      test_transmitter_visibility │   586,894,693ns │        1.6% │    2.39s │     4 │
│ test_first_triangles_hit_by_rays │   310,131,980ns │        4.9% │    2.98s │     9 │
│   test_compute_paths[exhaustive] │     6,087,314ns │        7.9% │    3.00s │   440 │
│       test_compute_paths[hybrid] │ 2,727,340,928ns │        0.0% │    2.73s │     1 │
│                  test_train_step │     1,443,286ns │        8.4% │    2.80s │ 1,686 │
└──────────────────────────────────┴─────────────────┴─────────────┴──────────┴───────┘
```

**After**

```
                                  Benchmark Results                                  
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃                        Benchmark ┃   Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│                test_image_method │      87,710ns │       18.9% │    2.89s │ 8,238 │
│                      test_fermat │  85,544,866ns │        2.6% │    2.75s │    31 │
│ test_rays_intersect_any_triangle │ 170,510,130ns │        8.9% │    2.44s │    13 │
│      test_transmitter_visibility │ 101,174,437ns │        3.6% │    2.88s │    27 │
│ test_first_triangles_hit_by_rays │ 268,347,521ns │        2.9% │    2.52s │     9 │
│   test_compute_paths[exhaustive] │   4,729,915ns │        7.5% │    2.94s │   540 │
│       test_compute_paths[hybrid] │ 633,988,360ns │        3.1% │    2.66s │     4 │
│                  test_train_step │   1,535,796ns │       16.0% │    3.00s │ 1,708 │
└──────────────────────────────────┴───────────────┴─────────────┴──────────┴───────┘
```

In the future, we might want to better fine-tune the default value for `batch_size` (here it was chosen to maximize the above performance).